+++ oxe-api/resource/form/add_form.py
@@ -1,0 +1,35 @@
+from flask_apispec import MethodResource
+from flask_apispec import use_kwargs, doc
+from flask_jwt_extended import jwt_required
+from flask_restful import Resource
+from webargs import fields
+
+from decorator.catch_exception import catch_exception
+from decorator.log_request import log_request
+from decorator.verify_admin_access import verify_admin_access
+
+
+class AddForm(MethodResource, Resource):
+
+    db = None
+
+    def __init__(self, db):
+        self.db = db
+
+    @log_request
+    @doc(tags=['form'],
+         description='Add a form',
+         responses={
+             "200": {},
+         })
+    @use_kwargs({
+        'name': fields.Str(),
+    })
+    @jwt_required
+    @verify_admin_access
+    @catch_exception
+    def post(self, **kwargs):
+
+        self.db.insert(kwargs, self.db.tables["Form"])
+
+        return "", "200 "