+++ oxe-api/resource/form/get_forms.py
@@ -1,0 +1,31 @@
+from flask_apispec import MethodResource
+from flask_apispec import doc
+from flask_jwt_extended import jwt_required
+from flask_restful import Resource
+
+from db.db import DB
+from decorator.catch_exception import catch_exception
+from decorator.log_request import log_request
+from decorator.verify_admin_access import verify_admin_access
+from utils.serializer import Serializer
+
+
+class GetForms(MethodResource, Resource):
+
+    def __init__(self, db: DB):
+        self.db = db
+
+    @log_request
+    @doc(tags=['form'],
+         description='Get the forms',
+         responses={
+             "200": {},
+         })
+    @jwt_required
+    @verify_admin_access
+    @catch_exception
+    def get(self):
+
+        data = Serializer.serialize(self.db.get(self.db.tables["Form"]), self.db.tables["Form"])
+
+        return data, "200 "